+++ Heyou/Classes/Elements/Element.swift
@@ -1,5 +1,3 @@
-import UIKit
-
 public protocol Element {
     func renderize() -> UIView
 }
