+++ Heyou/Classes/Elements/ImageElement.swift
@@ -1,5 +1,3 @@
-import Foundation
-
 public extension Heyou {
     struct Image: Element {
